BCC Capability & README Update
Updating README. Adding the option to add an extra BCC argument to the script (if needed for troubleshooting, for example). Also cleaning up the "Get-ADComputer" section just a tad to make sure that if a non-AD computer runs the script with no Clients List supplied, it will complain and NOT proceed.