Notification Filtering Improvements
Added needed fixes to the filtering mechanism for report creation. This well help to fight notification fatigue for administrators that are worried about missing a crucial piece within the normalcy of things like changing services. Updated README and scratching out a TODO from the list. Still need to tackle the slowness of the script.