Multiple Changes
I realized in my rush to build this that I've left a glaring issue (see #1 in the repo) which was that I'm executing all of the iterations on the remote machine instead of just fetching the info and doing it locally. Fixing that, and simultaneously modularizing the ability to add new data fields with a "TrackedValues" variable (addressing the other/final TODO on the README).