Added -NoMini & Started Tracking Module
Added a switch to the script to undo the new compression/mini-fication on the generated JSON files -- which has cut their sizes by almost 75%! -- in case admins want to be able to easily parse/read their reports. Started work on the filename tracking functionality that will measure counts of each given regex pattern in the specified directories (custom or predefined). This is not yet completed -- this is just an intermediate update before I get too lost. Also updated README and added more aesthetics to the console.