Fixed Null Pointer Exception & Online Tracking
At line 751 or so, there was a bug that was causing the invoke to fail out when the Contains method call was attempted on profilesMounted (if it was null). The report will now also be properly generated and tracked when the client comes back online and makes a change (between the offline check and the new check).