Added Debug Features & More TODOs
Added the Debug parameter, which sets the $DebugVerbosity level in the script, acting as a threshold for the verbosity of debug output. Also changed many of the original Client Monitor functions to have PowerShell standard verbs and formatting in their declarations. Trimmed a lot of fat, but added some in other places.