Adding Pre-Filtering for Repetitive Updates
Added a pre-filtering mechanism for installed applications that will suppress any further DisplayName or DisplayVersion updates for CHANGED INSTALLEDAPPS once they are initially detected and put into a notification digest, provided the script is set to automatically do so (off by default). What this effectively means is that you can define certain applications and versions that you no longer want to see updates for. This will be expanded once the "Examples" folder of the project docs is fleshed out.