Unsuppress Non-Version-Changing Installed Apps
The previous version of the monitor would properly only auto-track an installed application when the version changed, but changed apps that had NO DisplayName/DisplayVersion changes (and instead had InstallLocation or InstallDate changes, for example) were being suppressed from the notification because they matched the display name and then display version. I've built a requirement into the generation section that the prior version MUST NOT MATCH the new version in order for the item to be pre-filtered based on the changed version filtering mechanism.