License Change & New Update
Changed the license from MIT to GPLv3 and stamped the headers of all old and new source files. Yeah, this is basically an entirely new project at this point with how badly the old version has been deprecated. But it's still performing the same things, so it doesn't get its own space on my GitHub account. Take a peek at the commit and around the repo to see what's new. Also, check out the new README.