Updated Filename Tracking
Improving the performance of the per-profile filename tracking by taking the Profile variable check out of the foreach loop and moving it to just before. This helps the function and remote client by only requiring the profile-based filtering be run a single time, and not multiple times in the same piped command line.