PackageUserInformation Splice Fix
Fixed an issue where splicing the _prior suffix onto some objects where changes were detected was spitting out tons of errors with the PackageUserInformation_prior field being added more than once. This is because the script was adding PackageUserInformation to the TrackedValues array for StoreApps with every Get-ClientDeltas call (so per client). Move this += operation to BEFORE the foreach loop in the comparison section.