Added "Real Name" Translation
Added the ability to configure a real-name translation over HTTP or via a static Hashtable, which may allow notifications for each client to be more descriptive than simply their hostname. This also unlocks the ability to specify (indirectly) the use of an IP address or a hostname specifically in the generated report.