Slight Correction
Actually, the script will continue to force the use of the "Hostname" property as the client header field, and will include the "real name" on the same line if and only if it's defined. I may add styles or tags to this appended section (real name) at a later time, to differentiate it.