Offline/Online Propagation
Fixed an issue where a client becoming unreachable causes the previous report to be sent forward with both the IsOnline and IsInvokable properties in the Profile set to True. This could cause a continual loop of "offline to online" notifications for the client when Client Monitor is run while the targeted client is offline. Should be fixed now. Also added other minor fixes that are mostly aesthetic.