Small Offline-to-Online Bugfix
A bug was causing clients going from unreachable to reachable to lose the status changes. Thus the notifications would not include notifications for clients that went from an unreachable state to a reachable state. This is because the return variable from the remote invocation for gathering information populates both status variables to "False" as scaffolding, but they weren't being restored. Now they are. :]