Another Small Patch
After testing on my local computer, I realized that having 0 reachable clients (after all sessions drop for one reason or another) would require the Gathering phase to stop, so it can continue forward and report the dead sessions/rollbacks. Also, fixed an issue where only a single client in the $clients list wouldn't typecast to an ArrayList properly in two locations of ClientSessions.ps1. Also added forced nullification of the session variable, which may or may not be necessary, but doesn't seem to hurt any.