Better IP Alerts & Name Tracking Section
Fixed an issue where the ONLY delta for a client being an IP alert would cause the notification to omit the client header. Added a bit more console output. Fixed the notification section for filename tracking to finally sort itself by subcategory (i.e. Location) instead of Pattern, which was not a good choice for a table header (as it wasn't a unique property). Cleaned both the output itself AND the methods used to generate the desired output SUBSTANITALLY. I'm jumping the gun a bit on pushing this update so soon, but will watch this in more recent logs to see how it behaves.